Move ui folder inside components folder to better organize the code, update imports
