+++ src/components/Home/SearchSection/index.jsx
@@ -3,7 +3,7 @@ import { Typography } from 'antd'
 import AnimatedProvidersList from './AnimatedProvidersList'
 import { ErrorBoundary } from 'react-error-boundary'
 import { resetApplication } from '../../../common/utils'
-import { ErrorCard } from '../../../ui/Error/ErrorCard'
+import { ErrorCard } from '../../ui/Error/ErrorCard'
 import { useTranslation } from 'react-i18next'
 
 const { Title } = Typography
