+++ src/components/Home/Services/index.jsx
@@ -2,7 +2,7 @@ import { Card, Typography } from 'antd'
 import ServicesList from './ServicesList'
 import { ErrorBoundary } from 'react-error-boundary'
 import { resetApplication } from '../../../common/utils'
-import { ErrorCard } from '/src/ui/Error/ErrorCard'
+import { ErrorCard } from '../../ui/Error/ErrorCard'
 import { useTranslation } from 'react-i18next'
 
 const { Title } = Typography
