+++ src/components/Home/TopRatedDoctors/index.jsx
@@ -1,9 +1,9 @@
 import { Suspense, useState, useEffect, useMemo } from 'react'
 import { Pagination } from 'antd'
 import TopRatedDoctorCard from './TopRatedDoctorCard'
-import LoadingStatus from '../../../ui/Status/LoadingStatus'
+import LoadingStatus from '../../ui/Status/LoadingStatus'
 import TopRatedDoctorsCallToAction from './TopRatedDoctorsCallToAction'
-import { ErrorCard } from '../../../ui/Error/ErrorCard'
+import { ErrorCard } from '../../ui/Error/ErrorCard'
 import { ErrorBoundary } from 'react-error-boundary'
 import { resetApplication } from '../../../common/utils'
 import { fetchProviders } from '../../../lib/crud/fetchProviders'
