+++ src/components/Layout/index.jsx
@@ -2,7 +2,7 @@ import { ErrorBoundary } from 'react-error-boundary'
 import { Layout as AntDLayout } from 'antd'
 import Navbar from './Navbar'
 import Footer from './Footer'
-import { ErrorPage } from '../../ui/Error/ErrorPage'
+import { ErrorPage } from '../ui/Error/ErrorPage'
 import { resetApplication } from '../../common/utils'
 
 const { Content } = AntDLayout
