+++ src/components/Providers/AvailableSpots.jsx
@@ -10,7 +10,7 @@ import AvailabilityCards from './AvailabilityCard'
 import { getAvailableSpots } from '../../common/utils'
 import { ErrorBoundary } from 'react-error-boundary'
 import { resetApplication } from '../../common/utils'
-import { ErrorCard } from '../../ui/Error/ErrorCard'
+import { ErrorCard } from '../ui/Error/ErrorCard'
 import { useTranslation } from 'react-i18next'
 import { cap } from '../../common/utils'
