+++ src/components/Providers/ProviderCard.jsx
@@ -13,7 +13,7 @@ import {
 import Reviews from '../Reviews'
 import { ErrorBoundary } from 'react-error-boundary'
 import { resetApplication } from '../../common/utils'
-import { ErrorCard } from '../../ui/Error/ErrorCard'
+import { ErrorCard } from '../ui/Error/ErrorCard'
 import { useTranslation } from 'react-i18next'
 import { cap } from '../../common/utils'
