+++ src/components/Providers/ProviderModal.jsx
@@ -5,7 +5,7 @@ import ProviderCard from './ProviderCard'
 import AvailableTimes from './AvailableTimes'
 import SearchInput from '../SearchBar/SearchInput'
 import { ErrorBoundary } from 'react-error-boundary'
-import { ErrorCard } from '../../ui/Error/ErrorCard'
+import { ErrorCard } from '../ui/Error/ErrorCard'
 import { resetApplication } from '../../common/utils'
 import { useTranslation } from 'react-i18next'
 import { cap } from '../../common/utils'
