+++ src/components/Providers/index.jsx
@@ -7,10 +7,10 @@ import FewProvidersMessage from './FewProvidersMessage'
 import AvailableSpots from './AvailableSpots'
 import { SearchContext } from '../../providers/SearchProvider'
 import dayjs from 'dayjs'
-import LoadingStatus from '../../ui/Status/LoadingStatus'
+import LoadingStatus from '../ui/Status/LoadingStatus'
 import { toCamelCase } from '../../common/utils'
 import { ErrorBoundary } from 'react-error-boundary'
-import { ErrorCard } from '../../ui/Error/ErrorCard'
+import { ErrorCard } from '../ui/Error/ErrorCard'
 import { resetApplication } from '../../common/utils'
 import { fetchProviders } from '../../lib/crud/fetchProviders'
 import { useEffect } from 'react'
