+++ src/components/Reviews/index.jsx
@@ -4,7 +4,7 @@ import { Divider } from 'antd'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faStar } from '@fortawesome/free-solid-svg-icons'
 import { Suspense } from 'react'
-import LoadingStatus from '../../ui/Status/LoadingStatus'
+import LoadingStatus from '../ui/Status/LoadingStatus'
 import { useTranslation } from 'react-i18next'
 import { useState, useEffect } from 'react'
 import { fetchReviews } from '../../lib/crud/fetchReviews'
