+++ src/components/SearchBar/index.jsx
@@ -7,7 +7,7 @@ import { insurance } from '../../data/insuranceSearchList'
 import { location } from '../../data/locationSearchList'
 import { medicalConditions } from '../../data/medicalConditionsList'
 import { ErrorBoundary } from 'react-error-boundary'
-import { ErrorCard } from '../../ui/Error/ErrorCard'
+import { ErrorCard } from '../ui/Error/ErrorCard'
 import { resetApplication } from '../../common/utils'
 import { useTranslation } from 'react-i18next'
 import { cap } from '../../common/utils'
