+++ src/components/ui/Buttons/TranslationButton/index.jsx
@@ -4,7 +4,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCaretDown } from '@fortawesome/free-solid-svg-icons'
 import i18n from 'i18next'
 import { useTranslation } from 'react-i18next'
-import { cap } from '../../../common/utils'
+import { cap } from '../../../../common/utils'
 
 export const TranslationButton = () => {
   const { t } = useTranslation('global')
